chore(tests): split validate_tenant_id tests into test_tenant_utils
Co-locate the validate_tenant_id tests next to the implementation at
onyx/db/engine/tenant_utils, and leave only drop_schema tests in
test_schema_management.
